delete unused constant, rename threshold
