+++ src/main/java/frc/robot/subsystems/DriveBase.java
@@ -28,7 +28,7 @@ import frc.robot.Constants.DriveConstants;
 import frc.robot.Constants.PathConstants;
 import lobstah.stl.math.LobstahMath;
 import lobstah.stl.motorcontrol.LobstahDifferentialDrive;
-import frc.robot.EstimatedRobotPose;
+import frc.robot.photonvision.EstimatedRobotPose;
 
 /**
  * A subsystem that controls the drive train (aka chassis) on a robot.
